% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesample.R
\name{addeff}
\alias{addeff}
\title{Adding a new efficacy  variable.}
\usage{
addeff(ds, pv)
}
\arguments{
\item{ds}{input data set}

\item{pv}{response probabilities for initial nonresponders and initial responders.}
}
\value{
A data frame like input with an additional efficacy value.
}
\description{
\code{addeff} generates a second efficacy variable dependent on first.
}
\examples{
#Create a treatment set
ct<-onesample(eff=0.10,aeprob=cbind(c(0.965,0.020,0.015,0,0),
   c(0.95,0.03,0.02,0,0), c(0.98,0.01,0.01,0,0)))
ct$group<-0
ct<-addeff(ct,c(0.07,0.90))
tr<-onesample(eff=0.36,aeprob=cbind(c(0.945,0.030,0.020,0.005,0),
   c(0.90,0.05,0.04,0.01,0),c(0.90,0.05,0.04,00.01,0)))
tr$group<-1
tr<-addeff(tr,c(0.18,0.90))
all<-rbind(ct,tr)
}
