% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesample.R
\name{onesample}
\alias{onesample}
\title{Simulating correllated efficacy and severity data}
\usage{
onesample(
  nn = 200,
  eff = 0.4,
  aeprob = cbind(c(0.8, 0.05, 0.05, 0.05, 0.05), c(0.85, 0.05, 0.05, 0.03, 0.02),
    c(0.9, 0.05, 0.03, 0.02, 0.01), c(0.95, 0.02, 0.01, 0.01, 0.01)),
  alpha = 1,
  durexp = NULL,
  severrel = 0,
  override = FALSE
)
}
\arguments{
\item{nn}{The number of samples.}

\item{eff}{A cutoff creating a binary efficacy variable from a convolution of exponentials.}

\item{aeprob}{A table whose columns represent adverse events, and whose rows represent probabilities of various severity levels.}

\item{alpha}{a multiplier for the shared component between efficacy, severity, duration.}

\item{durexp}{Overall duration expectation}

\item{severrel}{a multiplier for the shared component between severity, duration.}

\item{override}{If true fixes expected turation to 8.}
}
\value{
A data frame with the above columns.
}
\description{
\code{onesample} generates a data frame containing one efficacy variable and a number of adverse event duration and severity variables.
}
\examples{
#An example with adverse event duration 8 days, low correlation among
#efficacy, severity, and duration.  alpha=1
aaa<-onesample(durexp=8)
#An example with adverse event duration 8 days, high correlation among
#efficacy, severity, and duration.
aaa<-onesample(durexp=8,alpha=7)
#An example with adverse event duration 8 days, no correlation among
#efficacy, severity, and duration, extra correlation between severity and
#duration.  Setting severrel=1 isn't supported yet.
aaa<-onesample(durexp=8,severrel=1,alpha=0)
}
